Test cases TC04 and TC08 are covered now.

+++ Music_Review_Application_Database/dbo/Tables/AlbumReview.sql
@@ -3,7 +3,7 @@
   [id] int PRIMARY KEY,
   [albumId] int not null,
   [username] nvarchar(30) not null,
-  [albumScore] decimal not null,
+  [albumScore] int not null,
   [albumReview] nvarchar(2500), 
     CONSTRAINT [FK_AlbumReview_ToAlbumTable] FOREIGN KEY ([albumId]) REFERENCES [Album]([id])
 )
